Extract package persistence from update release
This extracts a secondary collaborator that takes the output of
PackageProcessor as input to persist the results.

There may yet be tests in the UpdateRelease spec that are testing
PackagePersister behavior and would be better to move over. The
PackagePersister tests that were moved were only those that were
discovered because the existing UpdateRelease tests were setting mock
expectations on UpdateRelease itself to validate its behavior and thus
were difficult to leave where they were.
