+++ src/bosh-director/lib/bosh/director.rb
@@ -192,6 +192,7 @@ require 'bosh/director/jobs/cleanup_artifacts'
 require 'bosh/director/jobs/export_release'
 require 'bosh/director/jobs/update_deployment'
 require 'bosh/director/jobs/update_release'
+require 'bosh/director/jobs/update_release/package_persister'
 require 'bosh/director/jobs/update_release/package_processor'
 require 'bosh/director/jobs/update_stemcell'
 require 'bosh/director/jobs/fetch_logs'
